Move _get_new_logical_types_for_imputed_data to nullable type utils file
