+++ evalml/pipelines/components/transformers/imputers/target_imputer.py
@@ -8,10 +8,8 @@ from sklearn.impute import SimpleImputer as SkImputer
 from evalml.exceptions import ComponentNotYetFittedError
 from evalml.pipelines.components import ComponentBaseMeta
 from evalml.pipelines.components.transformers import Transformer
-from evalml.pipelines.components.transformers.imputers.utils import (
-    _get_new_logical_types_for_imputed_data,
-)
 from evalml.utils import infer_feature_types
+from evalml.utils.nullable_type_utils import _get_new_logical_types_for_imputed_data
 
 
 class TargetImputerMeta(ComponentBaseMeta):
